Development Completed
finally, I have managed to complete the development of THE CONTACT MANAGEMENT Application. This application stores its data in a JSON SERVER and it also allows the user to search for contacts, edit an already existing contact as well as create a new contact without data loss. in the development of this Application, I managed to use modern web development technologies and advanced React Js methodology.